Always return file content as binary data
Constructing XML files involves appending to lots of binaries, which is faster
if we use IO data and construct nested lists. When we encode to gzip, this gets
turned into a binary, which `ExAws.S3` accepts for upload.

However, when we disable gzip, we pass the IO data straight through. `ExAws.S3`
doesn't expect a file in IO data format, so chokes when calculating the
content-length. Instead, lets always convert to a binary even if we're not
gzipping, so implementations of `Sitemapper.Store`, don't have to handle both.